feat: add media upload service using Cloudinary

+++ Footbook.Infrastructure/Services/Implementations/FieldService.cs
@@ -1,7 +1,9 @@
+using CloudinaryDotNet.Actions;
 using FluentValidation;
 using Footbook.Core.DTOs.Requests.Field;
 using Footbook.Core.DTOs.Responses.Field;
 using Footbook.Data.Repositories.Interfaces;
+using Footbook.Infrastructure.ExternalServices.Cloudinary;
 using Footbook.Infrastructure.Helpers;
 using Footbook.Infrastructure.Services.Interfaces;
 
@@ -10,35 +12,46 @@ namespace Footbook.Infrastructure.Services.Implementations;
 public class FieldService : IFieldService
 {
     private readonly IFieldRepository _fieldRepository;
+    private readonly ICloudinaryService _cloudinaryService;
     private readonly IValidator<CreateFieldRequest> _createFieldValidator;
     private readonly IValidator<UpdateFieldRequest> _updateFieldValidator;
     
     public FieldService(
         IFieldRepository fieldRepository,
+        ICloudinaryService cloudinaryService,
         IValidator<CreateFieldRequest> createFieldValidator,
         IValidator<UpdateFieldRequest> updateFieldValidator)
     {
         _fieldRepository = fieldRepository;
+        _cloudinaryService = cloudinaryService;
         _createFieldValidator = createFieldValidator;
         _updateFieldValidator = updateFieldValidator;
     }
     
-    public async Task<CreateFieldResponse> CreateAsync(CreateFieldRequest request)
+    public async Task<FieldResponse> CreateAsync(CreateFieldRequest request)
     {
         await _createFieldValidator.ValidateAndThrowAsync(request);
         
         var field = request.MapToField();
+        
+        var image = request.Image;
+        if (image is not null)
+        {
+            var uploadResult = await _cloudinaryService.UploadImageAsync(image, "fields");
+            field.ImageUrl = uploadResult.SecureUrl.ToString();
+        }
+        
         var created = await _fieldRepository.CreateAsync(field);
-        return created.MapToCreateFieldResponse();
+        return created.MapToFieldResponse();
     }
     
-    public async Task<IEnumerable<CreateFieldResponse>> GetAllAsync()
+    public async Task<IEnumerable<FieldResponse>> GetAllAsync()
     {
         var list = await _fieldRepository.GetAllAsync();
-        return list.Select(f => f.MapToCreateFieldResponse());
+        return list.Select(f => f.MapToFieldResponse());
     }
     
-    public async Task<CreateFieldResponse> GetByIdAsync(Guid id)
+    public async Task<FieldResponse> GetByIdAsync(Guid id)
     {
         var field = await _fieldRepository.GetByIdAsync(id);
         
@@ -47,20 +60,46 @@ public class FieldService : IFieldService
             throw new KeyNotFoundException("Field not found.");
         }
 
-        return field.MapToCreateFieldResponse();
+        return field.MapToFieldResponse();
     }
     
-    public async Task<UpdateFieldResponse> UpdateAsync(Guid id, UpdateFieldRequest request)
+    public async Task<FieldResponse> UpdateAsync(Guid id, UpdateFieldRequest request)
     {
         await _updateFieldValidator.ValidateAndThrowAsync(request);
         
         var field = request.MapToField(id);
+        
+        var image = request.Image;
+        if (image is not null)
+        {
+            var imageUrl = field.ImageUrl;
+            if (!string.IsNullOrEmpty(imageUrl))
+            {
+                await _cloudinaryService.DeleteFileAsync(imageUrl, ResourceType.Image);
+            }
+            var uploadResult = await _cloudinaryService.UploadImageAsync(image, "fields");
+            field.ImageUrl = uploadResult.SecureUrl.ToString();
+        }
+        
         var updated = await _fieldRepository.UpdateAsync(field);
-        return updated.MapToUpdateFieldResponse();
+        return updated.MapToFieldResponse();
     }
     
     public async Task DeleteAsync(Guid id)
     {
+        var field = await _fieldRepository.GetByIdAsync(id);
+        
+        if (field is null)
+        {
+            throw new KeyNotFoundException("Field not found.");
+        }
+        
+        var imageUrl = field.ImageUrl;
+        if (!string.IsNullOrEmpty(imageUrl))
+        {
+            await _cloudinaryService.DeleteFileAsync(imageUrl, ResourceType.Image);
+        }
+        
         await _fieldRepository.DeleteAsync(id);
     }
 }
